backend/pyproject.toml
[project]
name = "livekit-meeting-intelligence"
version = "0.1.0"
description = "Realtime meeting transcription and intelligence with LiveKit"
requires-python = ">=3.10"
dependencies = [
    "livekit>=0.16.0",
    "livekit-agents>=0.8.0",
    "livekit-plugins-silero>=0.6.0",
    "mlx-whisper>=0.3.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "api"]
